Update hibernate configurations and refine code in tests
Various updates have been made to the hibernate configurations by clearing the password field in the hibernate.cfg.xml files for better security. Additionally, for the test classes, changes were also made to improve readability and efficiency. This includes modifications to the lambda function in QueriesTest.java and alterations to variable names and querying mechanisms in CriteriaTest.java.