Update dependencies and refactor tests and models
This commit updates multiple dependencies in the pom files, including junit, hibernate-core and mysql-connector-j. It also includes several corrections and refactoring in the test and model classes. Particularly, it made changes in the Company and Employee models in the hibernate-relationships-complex project, and to the ManyToManyTest case. In addition, it added a toString method to the SmartPhone model in the hibernate-naming-strategy project.