Update dependencies and modify print statements in tests
The dependencies for junit-jupiter, hibernate-core, and mysql-connector-java were updated in pom.xml. Maven Central repository was added to better manage dependencies. In the tests, a print statement was added in FetchingAllTest for better debugging and improved readability. The FetchingOneTest was also updated for better error handling with regards to LazyInitializationException. Password in hibernate connection was removed in hibernate.cfg.xml.